Fix classes being loaded multiple times
Previously, there was a bug with how classes were loaded when a class
loader recursively requested resolution: these classes were always
loaded from the class loaders, even if they were already loaded. This
could result in a class being loaded multiple times if a class reference
was found to an array of that class.