Implement method summarization
In order to make whole-program analysis (e.g. eliminating dead methods)
easier, a new form of method summarization has been implemented.
Effectively, this gives a brief summary of what large scale actions a
method may perform when called. For instance, it includes a list of
classes that might be initialized and other virtual methods that might
be called.