Implement recursive override resolution
Previously, overrides were only resolved 1 level. So, if A.foo overrode
B.foo which itself overrode C.foo, A.foo would not be marked as an
override of C.foo. For analysis purposes, it is necessary to be able to
find these overrides, so support for doing so has now been added.