Implement a very basic whole-program dead code analysis
Due to the way that JVMs are expected to perform lazy evaluation,
eagerly loading all classes can lead to many more classes being loaded
than are actually necessary. To combat this, a new pass has been added
that uses the previously gathered method summaries to attempt to find
dead classes and methods that don't need to be generated.