Add special ClassId for java/lang/Class
Due to the fact that objects of type java/lang/Class are only ever
expected to be created internally by the VM, we need some way to easily
refer to this class at compile time. Thus, this class has been added to
the list of bootstrap classes with a reserved ClassId.