Switch to using absolute offsets for branch bytecodes
Previously, the bytecode parser would provide relative offsets for any
branch opcodes. However, these are relative to the *start* of the branch
bytecode. As a result, consumers of this information would need to keep
track of the offset of the current bytecode themselves. To alleviate
this need, the relative address computation will now be performed in the
bytecode parser itself and absolute offsets will be provided.