Add infrastructure for a compile-time heap
In order to be able to effectively perform certain analyses, it is
necessary to be able to model the initial runtime state of the heap,
including java/lang/Class objects, constant string objects, and static
field values. To allow this, a new module has been added that contains
logic for modelling a Java heap at compile time.