Add handling for array classes loaded by anewarray
Previously, the class resolver assumed that the only classes that needed
to be loaded were those that were directly referred to by class constant
pool entries. However, it turns out that there is an additional way to
require a class without referring to it directly: the anewarray opcode
only requires that the element class be referred to directly. This is a
problem since array classes loaded by anewarray opcodes still need
java/lang/Class objects. To combat this, the resolver has been updated
to explicitly look for anewarray opcodes and load the necessary classes
at resolution time.