Add experimental support for running clinits at compile-time
In order to reduce runtime overhead, it is best to try and ensure that
all classes are initialized ahead of time. To allow this, an
experimental feature has been added where <clinit> methods of classes
that will need to be initialized will be compile-time evaluated using a
very basic interpreter if possible.