Include a ClassId for the type of a field
Previously, the Field structure did not include any resolution
information about what type of data the field held. While it was
possible to infer this from the TypeDescriptor by examining the current
ClassEnvironment, this defeats the point of having a separate resolution
phase. To avoid this, the Field struct now contains a ClassId for the
type of the field.