Properly mark classes referred to by ldc as needing clinit
Previously, classes whose only use was having their class object loaded
by using the ldc or ldc2 bytecodes would not be correctly marked as
requiring class initialization. This would result in the java/lang/Class
object for that class never being created, which would cause a crash
later when that bytecode was visited.