Fix calling java/lang/Object methods on arrays
Previously, the code for handling virtual method dispatch always assumed
that the receiver object would be a user class. However, it is possible
to call methods defined on java/lang/Object on an array class, in which
cause the java/lang/Object version of the method should be called.

This issue has been fixed by treating virtual dispatch as special
dispatch when the receiver object is not a user class.