Fix heap not rolling back after failed <clinit>
Due to a missing call to JavaStaticHeap::rollback, the state of the heap
would not be rolled back after failing to statically run a <clinit>
method. This would result in some changes made by the <clinit> method
persisting while others would not, leaving the heap in an indeterminate
state. This has been corrected.