Implement a basic IL
In order to emit machine code for a Java program, it will at some point
be necessary to output LLVM IR corresponding to each Java method and use
LLVM to optimize and emit an ELF executable. However, there are a number
of optimizations that LLVM may not be able to perform since LLVM does
not have the necessary type information, heap information, etc. to be
able to do so. In order to allow such optimizations to be added in the
future, the Java bytecode will instead be transformed to a higher-level
IL, which will be optimized internally before being lowered to LLVM IR.

Currently, a very basic IL has been added with the ability to model
enough basic straight-line code to compile a very simple "Hello World"
program using a custom native for printing. Further bytecode support and
optimizations will be added later, but for now this IL will simply be
used as-is to emit LLVM IR to allow the prototype of Mocha to be
completed quickly.