Add StaticInterpretError::OutOfMemory
Previously, if the static heap became full, the error returned would be
StaticInterpretError::WouldThrowException. However, this isn't strictly
correct since it's not guaranteed that an exception would be thrown at
runtime. This is due to the fact that the static heap often has a much
smaller size than the dynamic heap would, so running the failing
<clinit> method at runtime may actually succeed. As a result, a new
variant of StaticInterpretError has been added to represent this.