Fix <clinit> order not being correct sometimes
Due to missing calls to try_run_clinit_without_checkpoint, there were
several actions that could be taken without running the <clinit> of
another class when it should be necessry to do so. Several of these are
now fixed and the <clinit> of the class being used should be correctly
run.