Fix wrong dummy implementation of Class.forName
The previous dummy implementation of Class.forName would simply always
return null as the class. However, because the documentation for
Class.forName says that it should either return a class object or throw
an exception, code written to use that function was not expecting it to
return null. The dummy function has thus been switched to throw
ClassNotFoundException instead.