Add constants for object header size and Class object size
In the layout module, there are references to the size of the header of
a java/lang/Object and to the specific size of a java/lang/Class object,
as they are important to how fields are laid out. These have now been
put into constants so they are easier to understand.