Store return type ClassId on call instructions
Previously, call instructions would only store the MilType of their
return value. However, one goal of MIL is to represent enough
information to perform basic type deduction without having to peek at
other methods. In order to do this, it is necessary to know the actual
ClassId of the return value of a called function. This information is
now stored in the IL, and the MilType of the return value can be
inferred from the ClassId.