Generate IL thunks for native methods
Previously, no IL would be generated for native methods. However, there
are a couple special cases that need to be handled when calling a
native. For instance, when calling a static native, the first argument
should be the java/lang/Class object for the class on which the method
was defined. This is now handled by generating a small amount of thunk
code for each native method which is in charge of performing the actual
native call.