Exclude <clinit> for System from running statically
As it turns out, the OpenJDK implementation of the java/lang/System
class requires special initialization to be performed. As a result,
static initialization of this class is likely not to be fully possible.

For now, the <clinit> method of this class has been explicitly excluded
from consideration when running static class initialization. Obviously,
this isn't ideal, but it's enough to prevent crashes during static class
initialization.