Add a dummy AccessController class
In a dynamic configuration, the java/security/AccessController class is
used to run privileged actions via JNI. However, in a static compilation
scenario where SecurityManagers are not supported, there is no need to
perform native calls to do this.

In order to reduce runtime overhead of JCL classes that use APIs on this
class, a dummy pure-Java implementation has been added.