Reorganize the directory structure
Previously, the directory struct was somewhat messy, with the compiler
being in the root directory and the JCL being in a subdirectory. This
will become an even bigger problem when a runtime crate is inevitably
added to perform GC and VM work.

In order to keep things clean, the compiler has been moved into its own
directory and all Java source files for the modified JCL have been moved
into jcl/src.