Replace sipush and bipush with iconst
Previously, the bytecode reader would return different variants for the
iconst, sipush, and bipush instructions. However, sipush and bipush are
identical to iconst other than the size of their operand in the
bytecode.

To simplify handling of the bytecode elsewhere, the bytecode reader has
now been changed to emit BytecodeInstruction::IConst with the correctly
extended operand for all three instructions.