Fix bugs related to field alignment
It turns out that the previous logic for field layouts had a couple bugs
related to field alignment. For starters, the code for aligning static
fields was using the offset of the next *instance* field at one point,
which was causing static fields to be misaligned. On top of this, the
code for inserting padding fields in LLVM bitcode was calculating the
negative padding size and was trying to insert an array of that size
into the object, causing LLVM to break in mysterious ways.

These bugs have now been corrected and field alignment should now once
again work correctly.