Add a very basic runtime library
In order to be able to execute any reasonable Java code, it will be
necessary to have a runtime support library. A new cdylib crate called
mocha-rt has been added which should provide runtime support for Java
programs compiled by Mocha. Currently, this crate only contains a couple
test methods for printing integers and strings which are being used to
test the functionality of the compiler.