Add a liveness check for unused Strings
Previously, java/lang/String objects were emitted for every string in
the constant pool of an initialized class. However, not all strings
emitted for this were used, as they might be in the constant pool for
use by methods that end up being dead.

In order to reduce the number of strings emitted in the final binary,
the liveness checker will now also keep track of which strings are used
by live methods.