Add IL instructions for getting/setting locals
Previously, locals were simply treated as containing an SSA register
with the actual value. While this is usually fine, in order to allow
debugging where variables can be modified, it will be necessary to
instead treat them as allocas in LLVM.

This has been accomplished by adding GetLocal and SetLocal instructions
to MIL. In the future, converting these to SSA will likely be handled by
an optimizer pass.