Fix inconsistent vtable offsets
Previously, the offset of the vtable pointer on an object to the actual
start of the vtable was not consistent between objects created
statically and those allocated dynamically.

Due to what appears to be a strange LLVM quirk, trying to store the
vtable pointer directly as a 32-bit integer results in invalid assembly
code that cannot be relocated. This is apparently corrected when adding
a non-zero offset to the vtable pointer, so instead of using the start
of the vtable pointer, an offset of 8 is added. However, neither the
code for loading vtables nor the code for reading vslots took this into
account, resulting in incorrect vtable reads.

This has all now been corrected, and the +8 offset is now correctly
handled by all parts of the code.