Fix bad virtual calls of methods without vslots
When calculating object layouts, methods which are never overridden are
not given virtual slot numbers, i.e. they are not included in the
vtable. However, the previous logic of the IL generator would still emit
CallVirtual instructions when calling methods without a vslot using an
invoke_virtual bytecode. This would later result in a crash in the LLVM
code generator due to an out-of-bounds vslot read.

The IL generator has now been instructed to emit Call instructions when
encountering an invoke_virtual of a method without a virtual slot.