Add better runtime type information
Previously, the only available runtime information about a type was what
its object size was, which was stored on the Class object itself.
Firstly, there is no reason that this runtime-critical information
should be stored on a Class object. Secondly, this lacked information
required to be able to recognize array types correctly.

This situation has been improved by moving the object size into the
vtable and adding a lot of additional information to the vtable,
including information about an array class's element type.