Fix missing element classes for arrays
Previously, it was possible to initialize the class for an array type
without initializing the class for its element type. This would result
in a linker error due to the array class's vtable referring to the
vtable of the nonexistent element class.

This has been corrected by forcing the element class to initialize when
initializing an array class.