Fix wrong block boundaries during ilgen
Previously, there were ways to confuse the IL generator into inserting
an edge with the wrong bytecode as the predecessor block. This was due
to the fact that new jump targets could split what previously appeared
to be basic blocks, causing the originally added edge to refer to the
old start of the block.

This has now been corrected by first scanning the bytecode for branch
targets before scanning again to add the edges.