Fix exception throwing in LLVM backend
The previous handling of the stub exception throwing implementation in
the LLVM backend was using the default basic block terminator behaviour
of adding a branch to the next block. However, this caused a crash if
there was no next block, since the throw instruction was the last
instruction.

A better option is to place an 'unreachable' instruction after the call
to mocha_throw. This is a basic block terminating LLVM instruction which
tells LLVM that the call to mocha_throw can never return.