Fix incorrect shift directions in static interpreter
The previously implementations of the ishr, iushr, and ishl opcodes in
the static interpreter were shifting in the wrong directions. This
resulted in confusing and erratic behaviour, often resulting in a
<clinit> method failing due to an exception and thus forcing it to run
at runtime. Worse, it could result in the static <clinit> appearing to
have worked, but placing a wrong result somewhere in the static heap.

This has now been corrected and these opcodes should now work correctly
when run at compile time.