Fix issues with JumpIf for constant arguments
Previously, attempting to generate LLVM bitcode for a JumpIf instruction
where both of the operands were constants would result in incorrect LLVM
IR being generated. This occurred due to the fact that calling
LLVMBuildICmp did not emit an actual instruction in this case, meaning
that trying to find the compare instruction when generating the
conditional branch would instead find an unrelated instruction.

This has now been corrected by explicitly saving the LLVMValueRef of the
branch condition rather than trying to use the last instruction in the
basic block.