Fix crash when generating LLVM IR for get_static
Previously, the code for generating LLVM IR for a get_static IL
instruction would unnecessarily attempt to get the class object for the
type of the field it was trying to load. In the case of a static field
which is always null, it is perfectly valid for the class corresponding
to the type of the static field to not require initialization. This
would cause the codegen to panic due to the unnecessary lookup. Removal
of that lookup should correct this problem.