Stabilize order of function generation in LLVM codegen
Previously, the LLVM codegen would generate IR for functions in a random
order, since it used an iterator over a HashMap. This could result in a
single panic affecting generation of multiple functions appearing in
inconsistent locations. For ease of debugging, the LLVM codegen has been
modified to use a consistent order when generating functions.