Add null checks during ilgen
Previously, Mocha considered anything that would throw an NPE to be
undefined behaviour. However, the Java spec requires correct handling of
NPEs and many parts of the JCL rely upon this behaviour. While
exceptions are not yet fully supported, code has been added to the ilgen
to initiate a panic in any case where an NPE would normally be thrown.