Fix invalid calls and field loads/stores being generated
It's possible for invoke* bytecodes to refer to methods that cannot
actually ever be called if the receiver object of the call is always
null. Naively generating IL for these calls may cause issues if liveness
analysis detected that the target methods can never actually be called,
since the target method will never be generated nor put into the
itable/vtable.

This has now been corrected by adding a special lowering pass after
ilgen that looks for these invalid calls and will replace them with an
"unreachable" instruction. While this could be implemented as an
optimization, since it's mandatory to do this to get correct behaviour
from the codegen, it's better left separate.

The same treatment has also been given to putfield and getfield, which
can encounter the same issues.