Improve readability of Debug impl for FlowGraph<T>
Previously, the derived Debug impl for FlowGraph<T> produced output that
was very hard to understand. This was an issue when using the dbg! macro
during debugging, especially since the special IDs for exit and entry
nodes look very similar.

To combat this, a custom Debug impl has been added that should produce
more readable output.