Fix bad fcmpl/fcmpg implementation in the static interpreter
The previous implementation of the fcmpl and fcmpg bytecodes in the
static interpreter was incorrect. The static interpreter's version would
perform either an x < y or x > y comparison and push either 0 or 1.
However, the specification says that these opcodes should perform a
three-way compare and push either -1, 0, or 1. The static interpreter
has been corrected to implement this correctly.