Implement some basic class introspection
In order to support certain very basic operations, it's necessary for
certain introspection methods on java/lang/Class to be implemented. This
includes isArray() and isPrimitive().

Additionally, the java/lang/reflect/Array.newArray(...) method is now
supported in the static interpreter. This method is often used when
copying arrays, so it's very important that it be implemented during
class initialization.