Make MilRegisterMap::info private
Previously, the MilRegisterMap::info field was public and was often
manipulated directly. This was a problem for code that wanted to handle
MilRegister::VOID correctly, since that register is not part of this
map. To make things easier to maintain, this field is now public and
some helper methods have been added to MilRegisterMap for querying and
manipulating it.