Add stubs for instanceof/checkcast to IL generator
Stub implementations of the instanceof and checkcast bytecodes have now
been added to the IL generator. These implementations simply assume that
the object is always of the desired type. Obviously these will need to
be replaced later for correctness, but they're sufficient to get things
to compile for testing purposes.