Add fake String.intern() to static interpreter
A fake implementation of the String.intern() method that actually does
nothing has been added to the static interpreter. This is necessary
since the java/lang/StringBuffer.<clinit> method uses String.intern()
during class initialization.

Later, this should be replaced by a real implementation, but this works
well enough for the time being.