Add support for Object.clone on arrays
Previously, the Object.clone() method was not supported due to lacking
VM support. However, the JCL uses Object.clone() on arrays in several
places to create copies of arrays. In order to support this
functionality, preliminary support has been added for calling this
method on arrays. Full support for object cloning will be added at a
later date.