Implement basic field-based reflection
A very basic implementation of reflection on fields has now been
implemented. This includes a number of new Unsafe methods implemented
both in the static interpreter and at runtime, as well as an
implementation of Class.getDeclaredField(...) that works in the static
interpreter. The ability to get fields at runtime is not yet
implemented, since that would require persisting the raw field data for
the VM to use later.