Add proper setup for System.out and System.err
Previously, the System.err and System.out fields were always being set
to null, since the compiler was not yet ready to handle all the parts of
the JCL required for PrintStream to work correctly. However, all these
necessary dependencies now work correctly, so this functionality can now
be enabled.