Fix dead block elimination of blocks with edges to themselves
Previously, there was a bug in dead block elimination when trying to
remove blocks which had edges to themselves. Trying to remove CFG nodes
for these blocks would result in a panic due to trying to look up the
block's CFG node after it was removed. This has now been corrected and
blocks with edges to themselves can now be successfully removed.