Emit a table of builtin classes
Previously, builtin classes were only available during static
compilation. This is an issue since there are circumstances under which
the runtime library may need access to the class pointer for a builtin
class that isn't directly available through an existing argument. In
order to allow such lookups in the future, a new table will be emitted
in the binary that contains the address of each builtin class.