Add runtime init and shutdown routines
In order to be able to bootstrap a full JVM, it'll be necessary to run
certain tasks in the runtime library before initialization and prior to
JVM shutdown. To allow this to occur, two new functions have been added
to the runtime called mocha_init and mocha_shutdown that are executed
immediately before and immediately after execution of the Java main
method.