Implement devirtualization if receiver type is known exactly
In some cases (e.g. when an object is allocated within the current
method), it is possible to know the exact type of an object. Class
constraint analysis has been taught to track this information and
devirtualize virtual calls when possible if the type of the receiver
object is known exactly.