Initialize superclasses before their subclasses
Previously, there was no code to ensure that the <clinit> of a
superclass was run before that of any of its subclasses. While this is
usually fine as it will be initialized if the subclass attempts to use
anything from it, it does technically violate the specification to not
run the superclass's <clinit> first.