Suppress optimization logging when not passing --verbose
Previously, the optimizer would log its messages unconditionally. While
this was fine for smaller-scale tests, it quickly becomes an issue with
slightly larger projects. Not only is this unnecessary noise that the
user generally doesn't care about, it also causes a significant
performance hit due to the amount of string manipulation and system
calls requires. This has been corrected and the optimizer will now only
log if the --verbose option is specified.