Fix crash on invokespecial/invokestatic of interface method
Previously, the IL generator did not correctly implement the
invokespecial and invokestatic bytecodes with respect to calling methods
defined on interfaces. An invokespecial of an interface method can occur
when the user explicitly calls the superclass implementation of an
interface method that has a default implementation. An invokestatic of
an interface method can occur when said interface declares static
methods. In these cases, an InterfaceMethodref constant pool entry is
used instead of a normal Methodref constant pool entry, which the IL
generator failed to handle.