Implement tableswitch/lookupswitch via if-else chains
A basic implementation of the tableswitch and lookupswitch bytecodes has
been added to the IL generator that uses a chain of comparisons. While
it would be better to have dedicated opcodes for these, in the interests
of not complicating things for the moment, this has been left as a
future work item.