Prevent calling add_incoming on non-phi LLVM values
Previously, it was possible to call the add_incoming method on any
LLVMValue. This could result in undefined behaviour if called non a
non-phi node. To prevent this, a new LLVMPhiValue type has been added
which has this method and thus statically prevents calling add_incoming
on anything not resulting from the LLVMBuilder::build_phi method.