Allow gdb to skip parameter loading
Previously, gdb would only skip the part of the method prologue that was
generated by LLVM itself. This was an issue since the MIL code makes
loading the method parameters and storing them into their respective
local slots explicit. As a result, when stopping at function entry, the
value printed for these parameters would be incorrect.

To fix this, these prologue instructions are given a bytecode of !0,
which indicates to the codegen that they should not be given a debug
location. LLVM will then mark the end of the prologue as being the first
instruction having a debug location, automatically causing gdb to skip
the problematic instructions and start showing the correct values for
parameters immediately upon function entry.