Ensure that virtual stack phi nodes get correct bytecode
Previously, when generating phi nodes for stack merges in the IL
generator, the MilBuilder's bytecode would be set to whatever it
happened to be during the previous block's generation. This could cause
strange jumping in debuginfo. This has been corrected by ensuring that
all phi nodes are generated with the bytecode of the first instruction
in the block.