Use MilOperands for virtual stacks during ilgen
Previously, the virtual stack used during ilgen only supported storage
of MilRegisters. This resulted in a lot of unnecessary copy instructions
whenever constants were loaded that needed to be removed by the
optimizer. It is trivial (and in some cases actually easier) to use a
stack of MilOperands instead, which avoids these unnecessary copies
altogether.