Split verbose option for different phases
Previously, the option for enabling verbose logging would enable logging
for all compiler phases at once. However, this often makes it more
difficult to find information that the user is looking for, as enabling
logging would cause far too much output. In order to make it easier to
work with the verbose logging, the verbose option now accepts a list of
phases to log.