Ensure native functions are only declared once by LLVM codegen
Previously, if multiple call_native instructions appeared referencing
the same function, the LLVM codegen would attempt to declare that
function once each time. This causes LLVM to append a counter to the end
of the name to avoid naming conflicts, resulting in the linker being
unable to find the native function for all but one of the callsites.
This has been fixed by keeping a mapping around and ensuring that each
native function is only declared once.