Fix shared return blocks not having phi nodes updated
Previously, the inliner always assumed that the return block (the one
following a call which is being inlined) is not jumped to by any other
blocks. This assumption is incorrect, as the return value may be ignored
or only used as part of phi nodes on the return block. The inliner has
now been fixed to handle this case correctly.