Add strict-versioning to llvm-sys dependency
As it turns out, the compiler is using features of the LLVM C API that
are considered unstable and commonly change between LLVM versions. This
is a problem, since llvm-sys will use the system version of LLVM by
default even if it doesn't exactly match the version of LLVM expected.
This can result in segfaults due to mismatched parameters between the
compiler and the LLVM C library. To prevent this from happening, the
strict-versioning feature has been added to the compiler's llvm-sys
dependency, which causes llvm-sys to fail to build if the version of
LLVM found does not exactly match what is expected.