Cache the Cargo registry/index and target files in CI
Previously, the GitHub Actions workflow used for testing the compiler
and runtime would not cache anything between runs. This was inefficient,
as it would result in dependencies being fetched and built on every run.
The workflow has now been updated to cache the necessary files between
runs to speed this up.