Refactor constant folding
Previously, unary and binary operation constant folding would first
match on the types of operands and only then on the operation being
performed. This had the unfortunate side effect that adding a new
operation and forgetting to add a constant folding implementation for it
would result in a runtime panic if that operation was encountered with
constants.

To rectify this, the constant folder now matches on the operation
*before* matching on the operands. As a result, the match on operations
is now exhaustive and forgetting to add a new operation to constant
folding becomes a compile-time error.