Fix incorrect override detection for java/lang/Object methods
Due to a quirk of how all interfaces are technically said to be a
subclass of java/lang/Object, it was possible for interface methods to
be listed as virtual overrides of java/lang/Object methods. This is
clearly incorrect and can lead to strange behaviour later down the line.
For instance, it's possible that the java/lang/Object method itself to
be used to implement the relevant interface method. In this instance,
the Object method would be detected as overriding itself.