Add a more efficient way to lookup MIL instructions by register
In many cases, it's necessary to be able to efficiently perform lookups
of MIL instructions given a register. For instance, this is needed to be
able to perform simplifications such as ((x + 1) + 2) => (x + 3). The
current representation of MIL does not make this particularly easy, as
it is not possible to borrow instructions from one block mutably to
perform such transformations while simultaneously keeping borrows into
other blocks in order to perform such lookups, since it's technically
possible for the same block ID to appear twice in block_order.

Up to now, this has been handled by splitting such optimizations into
two passes: one to examine the MIL and prepare to make changes, and one
to actually apply the changes to the MIL instructions. However, this
isn't particularly efficient and is definitely not as ergonomic as being
able to perform analysis and transformation in one pass.

In order to handle this, a new API has been added that uses some unsafe
code to allow peeking at the instruction defining a register while
simultaneously iterating over the MIL instructions mutably. This API
allows modifying instructions and inserting new instructions, though it
does not allow adding/removing/rearranging blocks or deleting
instructions.