Replace constant/copy folding with "instruction simplification"
Previously, there was a dedicated optimization pass for performing
constant folding and copy propagation. This pass was very limited in
what it could do, as it was only able to look at each instruction
individually. This meant that simple transformations like turning
((x + 1) + 2) into (x + 3) couldn't be done during this pass.

To allow better optimization, a new "instruction simplification"
optimization has been added to replace the old opt. This optimization
uses the new MilFlatRepr API to be able to peek into other instructions
while iterating the IL. This allows much more powerful optimizations
than before and allows many optimizations to be performed in fewer
passes of the IL.