Fix duplicate phi entries after block merging
It turns out that under certain conditions, it was possible for block
merging to create phi nodes with duplicate entries. This would occur if
a block was being merged into its next block due to being empty, but had
duplicate predecessors. This would result in the computed value of
nonoverlap_new_incoming containing these duplicate predecessors, which
would then be included in the phi nodes.

This has been corrected by ensuring that nonoverlap_new_incoming is
deduplicated before being used.