Remove MilRegisterMap
Previously, every MIL function required a map defining the types of each
register. This was a common source of bugs, as forgetting to put a newly
allocated register into this map could cause panics much later. A better
solution is to define the types of registers at their point of use. This
allows the logic for determining the type of an operand to be
dramatically simplified. The validator will then check that the type
declared at each site where the register is used matches the type
declared at the site where the register was declared.