Fix block merging into loops incorrectly discarding phi nodes
Previously, it was possible to end up in a situation where the block
merger would attempt to merge the header block of a loop into the first
block of the loop. In this case, phi nodes from the header block can
legally be accessed in the loop since the header block will dominate all
loop blocks. However, the block merger always assumed that phi nodes on
the empty block could not be used outside of phi nodes since it assumed
that the blocks being merged would not have a dominance relation, and
would thus remove those phi nodes.

This issue has now been corrected and these phi nodes will be kept if
the blocks being merged have a dominance relationship, with newly added
predecessors using the phi node's value in the previous iteration.