Fix inlining of methods called immediately before loops
Similarly to the block merger bug that was just fixed, the inliner
assumed that the return block would not be dominated by the call block
if it had multiple predecessors. Therefore, it would not emit the phi
node corresponding to the return value, as it assumed it would be
invalid to access it directly. However, this is wrong if the next block
is the start of a loop with its predecessors being loop backedges.

In order to simplify things and correct this bug, the inliner will now
always create a new block to serve as the return block if the current
return block has multiple predecessors. This allows it to simply emit
the return value phi node and leave proper handling of this phi node to
block merging.