Make code generated for instanceof easier to simplify
Previously, code generated for instanceof checks could be quite
difficult for the instruction simplifier to recognize due to the select
being hidden behind a phi node. By moving the select to after the phi
node and using bool:false instead of int:0 on the phi node for null
references, this makes simplifier able to see through this pattern.