Explicitly specify the fallthrough of conditional jumps
Previously, conditional jumps would implicitly fall through to the next
block in lexical order. This inhibited certain optimizations, such as
folding znot/jc if the definition of the input to znot was not
invertible.

To make things easier, jc instructions now explicitly specify blocks to
jump to for both the true and false cases.