Improve select pattern recognition
Now that jc instructions can specify the target block when the condition
is false, select pattern recongnition's assumption that the true path
would always contain the merge block is no longer true. This
optimization has been updated so it will now work if the merge block is
on the false path.