Remove the MIL void type
Currently, MIL has a special $void register that it used to indicate
that an instruction does not actually yield a value, e.g. a void call.
However, this register was also sometimes used as a value with a "void"
type that was interpreted inconsistently. While this was usually
considered to be an invalid value entirely, it was being used as a
poison value on phi nodes created to represent Java locals. This could
result in MilOperands of register type not actually referring to
registers set by a prior instruction and introduced a lot of unnecessary
edge cases that weren't being handled correctly.

To make things more consistent, the void type has now been removed and
the $void register is now always illegal to use in a value context.
Instead, support for poison values has now been added that can be used
where $void was previously being used in phi nodes.