Make the validator ignore def-dom errors in unreachable blocks
Previously, it was possible for the validator to complain about the
definition of a register not dominating its uses in blocks that were
unreachable. Since these blocks are unreachable, it's safe to consider
all registers to have been defined. This even extends to registers which
were never defined at all.

This situation often occurs after ilgen when an unreachable instruction
causes a block to be truncated with an "unreachable" instruction. Future
blocks that are now unreachable would still refer to definitions from
instructions that were removed.