Fix incorrect phi predecessors in ilgen stack merging
During the introduction of explicit fallthroughs for conditional jumps,
a bug was introduced where the fallthrough target of a bytecode
conditional branch would get the wrong predecessor block applied to phi
nodes coming from stack merging.

This was caused by the taken and untaken edges technically originating
from different MIL blocks due to the addition of an explicit fallthrough
block. Normally, this block would be merged into the next one and the
problem would be hidden, but when adding validation after ilgen, these
started to get flagged as incorrect.

This problem also applies in a much worse way to switch bytecode
instructions, as they use a series of ifs to perform their checks. This
will be corrected shortly with some refactors.